# apps/api/internal/http/router.go
package http

import (
	"encoding/json"
	stdhttp "net/http"

	"github.com/go-chi/chi/v5"
	"github.com/jackc/pgx/v5/pgxpool"
	"github.com/namuh-eng/exponential/apps/api/internal/account"
	"github.com/namuh-eng/exponential/apps/api/internal/auth"
	"github.com/namuh-eng/exponential/apps/api/internal/comments"
	"github.com/namuh-eng/exponential/apps/api/internal/emojis"
	"github.com/namuh-eng/exponential/apps/api/internal/issues"
	"github.com/namuh-eng/exponential/apps/api/internal/labels"
	"github.com/namuh-eng/exponential/apps/api/internal/notifications"
	"github.com/namuh-eng/exponential/apps/api/internal/observability"
	"github.com/namuh-eng/exponential/apps/api/internal/projects"
	syncapi "github.com/namuh-eng/exponential/apps/api/internal/sync"
	"github.com/namuh-eng/exponential/apps/api/internal/teams"
	"github.com/namuh-eng/exponential/apps/api/internal/tokens"
	"github.com/namuh-eng/exponential/apps/api/internal/workspaces"
	"go.uber.org/zap"
)

// NewRouter wires API routes.
func NewRouter(logger *zap.Logger, db *pgxpool.Pool) stdhttp.Handler {
	metrics := &observability.Metrics{}
	r := chi.NewRouter()
	r.Use(observability.RequestLogger(logger, metrics))
	r.Get("/healthz", func(w stdhttp.ResponseWriter, r *stdhttp.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(stdhttp.StatusOK)
		if err := json.NewEncoder(w).Encode(map[string]string{"status": "ok"}); err != nil {
			logger.Error("write health response", zap.Error(err))
		}
	})
	r.Get("/metrics/red", func(w stdhttp.ResponseWriter, r *stdhttp.Request) {
		w.Header().Set("Content-Type", "application/json")
		w.WriteHeader(stdhttp.StatusOK)
		if err := json.NewEncoder(w).Encode(observability.Snapshot(metrics)); err != nil {
			logger.Error("write metrics response", zap.Error(err))
		}
	})

	authMiddleware := auth.Middleware{DB: db}
	commentsHandler := comments.Handler{DB: db}
	labelsHandler := labels.Handler{DB: db}
	r.Route("/v1", func(v1 chi.Router) {
		v1.Group(func(protected chi.Router) {
			protected.Use(authMiddleware.Require)
			protected.Post("/issues/{id}/comments", commentsHandler.CreateForIssue)
			protected.Post("/issues/{id}/reactions", commentsHandler.ToggleIssueReaction)
			protected.Delete("/issues/{id}/reactions", commentsHandler.DeleteIssueReaction)
			protected.Mount("/account", account.Handler{DB: db}.Routes())
			protected.Patch("/comments/{id}", commentsHandler.Update)
			protected.Mount("/custom-emojis", emojis.Handler{DB: db}.Routes())
			protected.Delete("/comments/{id}", commentsHandler.Delete)
			protected.Post("/comments/{id}/reactions", commentsHandler.ToggleCommentReaction)
			protected.Mount("/issues", issues.Handler{DB: db}.Routes())
			protected.Mount("/labels", labelsHandler.Routes())
			protected.Mount("/notifications", notifications.Handler{DB: db}.Routes())
			protected.Mount("/project-labels", labelsHandler.ProjectRoutes())
			protected.Mount("/personal-access-tokens", tokens.Handler{DB: db}.Routes())
			protected.Mount("/projects", projects.Handler{DB: db}.Routes())
			protected.Mount("/teams", teams.Handler{DB: db}.Routes())
			protected.Mount("/workspaces", workspaces.Handler{DB: db}.Routes())
			protected.Get("/sync/ws", syncapi.Handler{DB: db}.WebSocket)
		})
	})
	return r
}