# packages/proto/openapi.yaml
openapi: 3.1.0
info:
  title: Exponential Headless API
  version: 0.1.0
...
servers:
  - url: http://localhost:3016/v1
security:
  - bearerAuth: []
paths:
  /issues:
    get:
      operationId: listIssues
      summary: List issues in the authenticated workspace
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: team_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Issues page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createIssue
      summary: Create an issue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Created issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        default:
          $ref: '#/components/responses/Problem'
...
  /issues/{id}/comments:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: createIssueComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Created comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        default:
          $ref: '#/components/responses/Problem'
  /issues/{id}/reactions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: toggleIssueReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Issue reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteIssueReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Issue reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
...
  /comments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteComment
      responses:
        '200':
          description: Deleted comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        default:
          $ref: '#/components/responses/Problem'
  /comments/{id}/reactions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: toggleCommentReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Comment reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
  /custom-emojis:
    get:
      operationId: listCustomEmojis
      responses:
        '200':
          description: Custom emojis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEmojiListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createCustomEmoji
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomEmojiRequest'
      responses:
        '201':
          description: Created custom emoji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEmojiResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /labels:
    get:
      operationId: listLabels
      parameters:
        - name: scope
          in: query
          schema: { type: string, enum: [workspace, team, all], default: workspace }
        - name: teamId
          in: query
          schema: { type: string, format: uuid }
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Labels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createLabel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLabelRequest'
      responses:
        '201':
          description: Created label
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /project-labels:
    get:
      operationId: listProjectLabels
      responses:
        '200':
          description: Project labels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectLabelListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProjectLabel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectLabelRequest'
      responses:
        '201':
          description: Created project label
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectLabelResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /notifications:
    get:
      operationId: listNotifications
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /personal-access-tokens:
    get:
      operationId: listPersonalAccessTokens
      responses:
        '200':
          description: Personal access tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalAccessTokenListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createPersonalAccessToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonalAccessTokenRequest'
      responses:
        '201':
          description: Created personal access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePersonalAccessTokenResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /projects:
    get:
      operationId: listProjects
      summary: List projects in the authenticated workspace
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
...
  /teams:
    get:
      operationId: listTeams
      summary: List visible teams in the authenticated workspace
      responses:
        '200':
          description: Teams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createTeam
      summary: Create a team with default workflow states
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Created team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTeamResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /workspaces:
    get:
      operationId: listWorkspaces
      summary: List workspaces for the authenticated principal
      responses:
        '200':
          description: Workspace memberships
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WorkspaceMembership' }
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createWorkspace
      summary: Create a workspace with a default team and workflow states
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWorkspaceResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /sync/ws:
    get:
      operationId: watchSync
      summary: Open a workspace sync WebSocket and replay operations after a version
      parameters:
        - name: version
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: access_token
          in: query
          description: Browser/CLI WebSocket auth fallback when Authorization headers are unavailable.
          schema: { type: string }
      responses:
        '101':
          description: WebSocket protocol switch
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: lin_api or pat
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    Problem:
      description: RFC 7807 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    CommentUser:
      type: object
      required: [name]
      properties:
        name: { type: string }
        image: { type: [string, 'null'] }
    ReactionSummary:
      type: object
      required: [emoji, count]
      properties:
        emoji: { type: string }
        count: { type: integer }
        reacted: { type: boolean }
        reactedByMe: { type: boolean }
    ReactionSummaryList:
      type: array
      items: { $ref: '#/components/schemas/ReactionSummary' }
    Comment:
      type: object
      required: [id, body, issue_id, user_id, user, owned_by_me, can_edit, can_delete, reactions, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        issue_id: { type: string, format: uuid }
        user_id: { type: string }
        user: { $ref: '#/components/schemas/CommentUser' }
        owned_by_me: { type: boolean }
        can_edit: { type: boolean }
        can_delete: { type: boolean }
        reactions:
          type: array
          items: { $ref: '#/components/schemas/ReactionSummary' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
    UpdateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
    ReactionRequest:
      type: object
      required: [emoji]
      properties:
        emoji: { type: string, minLength: 1, maxLength: 50 }
    CustomEmoji:
      type: object
      required: [id, name, imageUrl, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        imageUrl: { type: string }
        createdAt: { type: string, format: date-time }
    CustomEmojiListResponse:
      type: object
      required: [emojis]
      properties:
        emojis:
          type: array
          items: { $ref: '#/components/schemas/CustomEmoji' }
    CustomEmojiResponse:
      type: object
      required: [emoji]
      properties:
        emoji: { $ref: '#/components/schemas/CustomEmoji' }
    CreateCustomEmojiRequest:
      type: object
      required: [name, imageUrl]
      properties:
        name: { type: string, minLength: 1, maxLength: 32 }
        imageUrl: { type: string }
    Label:
      type: object
      required: [id, name, color, scope, issueCount, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        description: { type: [string, 'null'] }
        parentLabelId: { type: [string, 'null'], format: uuid }
        teamId: { type: [string, 'null'], format: uuid }
        teamName: { type: [string, 'null'] }
        teamKey: { type: [string, 'null'] }
        scope: { type: string, enum: [workspace, team] }
        archivedAt: { type: [string, 'null'], format: date-time }
        issueCount: { type: integer }
        lastApplied: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LabelListResponse:
      type: object
      required: [labels]
      properties:
        labels:
          type: array
          items: { $ref: '#/components/schemas/Label' }
    LabelResponse:
      type: object
      required: [label]
      properties:
        label: { $ref: '#/components/schemas/Label' }
    CreateLabelRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        color: { type: string }
        description: { type: [string, 'null'] }
        parentLabelId: { type: [string, 'null'], format: uuid }
        teamId: { type: [string, 'null'], format: uuid }
...
    ProjectLabel:
      type: object
      required: [id, name, color, projectCount, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        description: { type: [string, 'null'] }
        projectCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectLabelListResponse:
      type: object
      required: [labels]
      properties:
        labels:
          type: array
          items: { $ref: '#/components/schemas/ProjectLabel' }
    ProjectLabelResponse:
      type: object
      required: [label]
      properties:
        label: { $ref: '#/components/schemas/ProjectLabel' }
    CreateProjectLabelRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        color: { type: string }
        description: { type: [string, 'null'] }
...
    Notification:
      type: object
      required: [id, type, actorName, issueIdentifier, issueTitle, issuePriority, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [assigned, mentioned, status_change, comment, duplicate] }
        actorName: { type: string }
        actorImage: { type: [string, 'null'] }
        issueIdentifier: { type: string }
        issueTitle: { type: string }
        issuePriority: { type: string }
        issueId: { type: [string, 'null'], format: uuid }
        readAt: { type: [string, 'null'], format: date-time }
        snoozedUntilAt: { type: [string, 'null'], format: date-time }
        unsnoozedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    NotificationListResponse:
      type: object
      required: [notifications, unreadCount]
      properties:
        notifications:
          type: array
          items: { $ref: '#/components/schemas/Notification' }
        unreadCount: { type: integer }
...
    PersonalAccessToken:
      type: object
      required: [id, name, token_prefix, scopes, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        token_prefix: { type: string }
        scopes:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
    PersonalAccessTokenListResponse:
      type: object
      required: [tokens]
      properties:
        tokens:
          type: array
          items: { $ref: '#/components/schemas/PersonalAccessToken' }
    CreatePersonalAccessTokenRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        scopes:
          type: array
          items: { type: string }
    CreatePersonalAccessTokenResponse:
      type: object
      required: [token, value]
      properties:
        token: { $ref: '#/components/schemas/PersonalAccessToken' }
        value: { type: string }
    ProjectStatus:
      type: string
      enum: [planned, started, paused, completed, canceled]
    ProjectPriority:
      type: string
      enum: [none, urgent, high, medium, low]
    ProjectTeam:
      type: object
      required: [id, key, name]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
    ProjectProgress:
      type: object
      required: [total, completed, percentage]
      properties:
        total: { type: integer }
        completed: { type: integer }
        percentage: { type: integer }
    Project:
      type: object
      required: [id, name, slug, status, priority, workspace_id, teams, progress, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'] }
        slug: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        priority: { $ref: '#/components/schemas/ProjectPriority' }
        lead_id: { type: [string, 'null'] }
        workspace_id: { type: string, format: uuid }
        start_date: { type: [string, 'null'], format: date-time }
        target_date: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        canceled_at: { type: [string, 'null'], format: date-time }
        teams:
          type: array
          items: { $ref: '#/components/schemas/ProjectTeam' }
        progress: { $ref: '#/components/schemas/ProjectProgress' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectListResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items: { $ref: '#/components/schemas/Project' }
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        description: { type: [string, 'null'] }
        icon: { type: [string, 'null'], maxLength: 10 }
        slug: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        priority: { $ref: '#/components/schemas/ProjectPriority' }
        lead_id: { type: [string, 'null'] }
        start_date: { type: [string, 'null'], format: date-time }
        target_date: { type: [string, 'null'], format: date-time }
        team_ids:
          type: array
          items: { type: string, format: uuid }
        team_keys:
          type: array
          items: { type: string }
...
    Team:
      type: object
      required: [id, name, key, isPrivate, issueCount, memberCount, currentUserIsMember, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key: { type: string }
        icon: { type: [string, 'null'] }
        isPrivate: { type: boolean }
        issueCount: { type: integer }
        memberCount: { type: integer }
        currentUserIsMember: { type: boolean }
        createdAt: { type: string, format: date-time }
        retiredAt: { type: [string, 'null'], format: date-time }
    TeamListResponse:
      type: object
      required: [workspaceId, viewerRole, canManageTeams, teams]
      properties:
        workspaceId: { type: string, format: uuid }
        viewerRole: { $ref: '#/components/schemas/WorkspaceRole' }
        canManageTeams: { type: boolean }
        teams:
          type: array
          items: { $ref: '#/components/schemas/Team' }
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        key: { type: string, maxLength: 10 }
        isPrivate: { type: boolean }
        icon: { type: string, maxLength: 16 }
    CreateTeamResponse:
      type: object
      required: [team]
      properties:
        team: { $ref: '#/components/schemas/Team' }
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, member, guest]
    WorkspaceMembership:
      type: object
      required: [workspaceId, role, workspaceName, workspaceSlug]
      properties:
        workspaceId: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/WorkspaceRole' }
        workspaceName: { type: string }
        workspaceSlug: { type: string }
...
    WorkspaceSummary:
      type: object
      required: [id, name, urlSlug, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        urlSlug: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamSummary:
      type: object
      required: [id, name, key, workspaceId]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key: { type: string }
        workspaceId: { type: string, format: uuid }
    CreateWorkspaceRequest:
      type: object
      required: [name, urlSlug]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        urlSlug: { type: string, minLength: 2, maxLength: 63 }
    CreateWorkspaceResponse:
      type: object
      required: [workspace, team]
      properties:
        workspace: { $ref: '#/components/schemas/WorkspaceSummary' }
        team: { $ref: '#/components/schemas/TeamSummary' }
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    IssuePriority:
      type: string
      enum: [none, urgent, high, medium, low]
    Issue:
      type: object
      required: [id, number, identifier, title, team_id, state_id, creator_id, priority, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        number: { type: integer }
        identifier: { type: string }
        title: { type: string }
        description: { type: [string, 'null'] }
        team_id: { type: string, format: uuid }
        state_id: { type: string, format: uuid }
        assignee_id: { type: [string, 'null'] }
        creator_id: { type: string }
        priority: { $ref: '#/components/schemas/IssuePriority' }
        estimate: { type: [number, 'null'] }
        parent_issue_id: { type: [string, 'null'], format: uuid }
        project_id: { type: [string, 'null'], format: uuid }
        project_milestone_id: { type: [string, 'null'], format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        due_date: { type: [string, 'null'], format: date-time }
        sort_order: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        archived_at: { type: [string, 'null'], format: date-time }
        canceled_at: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
    IssueListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
        next_cursor: { type: [string, 'null'] }
    CreateIssueRequest:
      type: object
      required: [title, team_id]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, 'null'] }
        team_id: { type: string, format: uuid }
        state_id: { type: [string, 'null'], format: uuid }
        priority: { $ref: '#/components/schemas/IssuePriority' }
        assignee_id: { type: [string, 'null'] }
        project_id: { type: [string, 'null'], format: uuid }
        project_milestone_id: { type: [string, 'null'], format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        parent_issue_id: { type: [string, 'null'], format: uuid }
        estimate: { type: [number, 'null'] }
        due_date: { type: [string, 'null'], format: date }
...